refactor: move DefaultIcons to IconStorage
